Add files via upload
People with windows can try out the modded version, hit TAB to switch between WASD/arrow-keys and WASD/mouse viewing modes. In the current version you can see the mouse cursor being moved back to the center position (which is strangely off by 9 pixels in x and 38 pixels in y). In the actual version we would probably want to hide the cursor in this mode and instead replace it with a cross-hair that is drawn directly onto the 2D HUD.

Future Stuff:
- Get it working on MAC
- Remove cursor in mouse-view mode
- Figure out what's causing the weird offset
- Move the camera framework entirely into the camera class